Simplify weather manager hierarchy
