+++ src/main/java/net/minestom/server/weather/manager/ForwardingWeatherManager.java
@@ -1,33 +1,41 @@
 package net.minestom.server.weather.manager;
 
+import net.minestom.server.MinecraftServer;
+import net.minestom.server.network.netty.packet.FramedPacket;
 import net.minestom.server.weather.Weather;
-import net.minestom.server.weather.container.WeatherContainer;
+import net.minestom.server.weather.WeatherContainer;
 import org.jetbrains.annotations.NotNull;
 
+import java.util.Collection;
 import java.util.Objects;
 import java.util.function.Supplier;
 
 /**
  * A weather manager that either has a weather or has the weather of it's parent.
  */
-public abstract class ForwardingWeatherManager extends WeatherManager {
-    private final Supplier<WeatherContainer> parent;
+public class ForwardingWeatherManager extends WeatherManager {
+    private final Supplier<? extends WeatherContainer> parent;
+    private final Supplier<Collection<? extends WeatherContainer>> children;
 
     /**
-     * Creates a new forwarding weather manager with a given parent.
+     * Creates a new forwarding weather manager with children.
+     *
+     * @param children the children
      */
-    public ForwardingWeatherManager(@NotNull WeatherContainer parent) {
-        super(null);
-        Objects.requireNonNull(parent, "parent");
-        this.parent = () -> parent;
+    public ForwardingWeatherManager(@NotNull Supplier<Collection<? extends WeatherContainer>> children) {
+        this(MinecraftServer::getGlobalWeatherManager, children);
     }
 
     /**
-     * Creates a new forwarding weather manager with a given parent.
+     * Creates a new forwarding weather manager with a parent and children.
+     *
+     * @param parent the parent
+     * @param children the children
      */
-    public ForwardingWeatherManager(@NotNull Supplier<WeatherContainer> parent) {
+    public ForwardingWeatherManager(@NotNull Supplier<? extends WeatherContainer> parent, @NotNull Supplier<Collection<? extends WeatherContainer>> children) {
         super(null);
         this.parent = Objects.requireNonNull(parent, "parent");
+        this.children = Objects.requireNonNull(children, "children");
     }
 
     /**
@@ -37,15 +45,6 @@ public abstract class ForwardingWeatherManager extends WeatherManager {
      */
     public @NotNull WeatherContainer getParentContainer() {
         return this.parent.get();
-    }
-
-    /**
-     * Checks if this weather manager has its own weather set.
-     *
-     * @return {@code true} if it does
-     */
-    public boolean hasWeather() {
-        return this.weather != null;
     }
 
     @Override
@@ -65,6 +64,15 @@ public abstract class ForwardingWeatherManager extends WeatherManager {
         if (this.weather != null) {
             this.sendWeatherPackets(WeatherManager.createWeatherPackets(this.weather, this.getParentContainer().getWeather()));
             this.weather = null;
+        }
+    }
+
+    @Override
+    protected void sendWeatherPackets(@NotNull Collection<FramedPacket> packets) {
+        if (!packets.isEmpty()) {
+            for (WeatherContainer weatherContainer : this.children.get()) {
+                weatherContainer.getWeatherManager().sendWeatherPackets(packets);
+            }
         }
     }
 }
